--- lib/models/firebase/firebase_user.dart
@@ -1,14 +1,0 @@
-
-
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:seeds/providers/services/firebase/firebase_database_map_keys.dart';
-
-class FirebaseUser {
-  final String? account;
-  final Timestamp? guardianRecoveryStarted;
-
-  FirebaseUser({this.account, this.guardianRecoveryStarted});
-
-  FirebaseUser.fromMap(Map<String, dynamic> data, String accountName)
-      : this(account: accountName, guardianRecoveryStarted: data[GUARDIAN_RECOVERY_STARTED_KEY]);
-}
