--- lib/providers/services/firebase/firebase_database_map_keys.dart
@@ -1,27 +1,0 @@
-
-
-const String GUARDIANS_COLLECTION_KEY = 'guardians';
-const String GUARDIANS_STATUS_KEY = 'status';
-const String GUARDIANS_DATE_CREATED_KEY = 'dateCreated';
-const String GUARDIANS_DATE_UPDATED_KEY = 'dateUpdated';
-const String GUARDIAN_CONTRACT_INITIALIZED = 'guardianContractInitialized';
-const String GUARDIAN_CONTRACT_INITIALIZED_DATE = 'guardianContractInitializedDate';
-const String GUARDIAN_CONTRACT_INITIALIZED_UPDATE_DATE = 'guardianContractInitializedUpdateDate';
-const String GUARDIAN_NOTIFICATION_KEY = 'guardianNotification';
-const String GUARDIAN_RECOVERY_STARTED_KEY = 'guardianRecoveryStarted';
-const String UID_KEY = 'uid';
-const String TYPE_KEY = 'type';
-const String RECOVERY_COLLECTION_KEY = 'recovery';
-const String RECOVERY_STARTED_DATE_KEY = 'recoveryStartedDate';
-const String RECOVERY_APPROVED_DATE_KEY = 'recoveryApprovedDate';
-
-const String PRODUCTS_COLLECTION_KEY = 'products';
-const String PRODUCT_NAME_KEY = 'productName';
-const String PRODUCT_PRICE_KEY = 'productPrice';
-const String PRODUCT_IMAGE_URL_KEY = 'imageUrl';
-const String PRODUCT_CREATED_DATE_KEY = 'createdDate';
-const String PRODUCT_UPDATED_DATE_KEY = 'updatedDate';
-const String PRODUCT_CURRENCY_KEY = 'productCurrency';
-const String PRODUCT_POSITION_KEY = 'position';
-
-const String PENDING_NOTIFICATIONS_KEY = 'pendingNotifications';
